+++ example/lib/widgets/insta_picker_interface.dart
@@ -1,3 +1,6 @@
+import 'dart:io';
+
+import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:insta_assets_picker/insta_assets_picker.dart';
 import 'package:insta_assets_picker_demo/main.dart';
@@ -27,6 +30,11 @@ mixin InstaPickerInterface on Widget {
   }
 
   AppBar get _appBar => AppBar(title: Text(description.fullLabel));
+
+  /// NOTE: Exception on android when playing video recorded from the camera
+  /// with [ResolutionPreset.max] after FFmpeg encoding
+  ResolutionPreset get cameraResolutionPreset =>
+      Platform.isAndroid ? ResolutionPreset.high : ResolutionPreset.max;
 
   Column pickerColumn({
     String? text,
@@ -79,10 +87,14 @@ mixin InstaPickerInterface on Widget {
   void pickAssets(BuildContext context, {required int maxAssets}) =>
       InstaAssetPicker.pickAssets(
         context,
-        title: description.fullLabel,
-        closeOnComplete: true,
+        pickerConfig: InstaAssetPickerConfig(
+          title: description.fullLabel,
+          closeOnComplete: true,
+          pickerTheme: getPickerTheme(context),
+          // skipCropOnComplete: true, // to test ffmpeg crop image
+          // previewThumbnailSize: const ThumbnailSize(240, 240), // to improve thumbnails speed in crop view
+        ),
         maxAssets: maxAssets,
-        pickerTheme: getPickerTheme(context),
         onCompleted: (Stream<InstaAssetsExportDetails> cropStream) {
           Navigator.push(
             context,