+++ lib/src/assets_picker.dart
@@ -3,7 +3,6 @@
 import 'package:flutter/material.dart';
 import 'package:insta_assets_picker/insta_assets_picker.dart';
 import 'package:insta_assets_picker/src/widget/insta_asset_picker_delegate.dart';
-import 'package:wechat_assets_picker/wechat_assets_picker.dart';
 
 const _kGridCount = 4;
 const _kInitializeDelayDuration = Duration(milliseconds: 250);
@@ -33,11 +32,111 @@ class InstaAssetCropDelegate {
   final List<double> cropRatios;
 }
 
+/// Configurations for the [InstaAssetPickerBuilder].
+class InstaAssetPickerConfig {
+  const InstaAssetPickerConfig({
+    /// [DefaultAssetPickerBuilderDelegate] config
+
+    this.gridCount = _kGridCount,
+    this.pickerTheme,
+    this.specialItemPosition,
+    this.specialItemBuilder,
+    this.loadingIndicatorBuilder,
+    this.selectPredicate,
+    this.limitedPermissionOverlayPredicate,
+    this.themeColor,
+    this.textDelegate,
+    this.gridThumbnailSize = defaultAssetGridPreviewSize,
+    this.previewThumbnailSize,
+    this.pathNameBuilder,
+
+    /// [InstaAssetPickerBuilder] config
+
+    this.title,
+    this.cropDelegate = const InstaAssetCropDelegate(),
+    this.closeOnComplete = false,
+    this.skipCropOnComplete = false,
+    this.actionsBuilder,
+  });
+
+  /* [DefaultAssetPickerBuilderDelegate] config */
+
+  /// Specifies the number of assets in the cross axis.
+  ///
+  /// Defaults to [_kGridCount], like instagram.
+  final int gridCount;
+
+  /// Specifies the theme to apply to the picker.
+  /// It is by default initialized with the `primaryColor` of the context theme.
+  final ThemeData? pickerTheme;
+
+  /// Set a special item in the picker with several positions.
+  /// Since the grid view is reversed, [SpecialItemPosition.prepend]
+  /// will be at the top and [SpecialItemPosition.append] at the bottom.
+  ///
+  /// Defaults to [SpecialItemPosition.none].
+  final SpecialItemPosition? specialItemPosition;
+
+  /// Specifies [Widget] for the the special item.
+  final SpecialItemBuilder<AssetPathEntity>? specialItemBuilder;
+
+  /// The loader indicator to display in the picker.
+  final LoadingIndicatorBuilder? loadingIndicatorBuilder;
+
+  /// Predicate whether an asset can be selected or unselected.
+  final AssetSelectPredicate<AssetEntity>? selectPredicate;
+
+  /// Specifies if the limited permission overlay should be displayed.
+  final LimitedPermissionOverlayPredicate? limitedPermissionOverlayPredicate;
+
+  /// Main color for the picker.
+  final Color? themeColor;
+
+  /// Specifies the language to apply to the picker.
+  ///
+  /// Default is the locale language from the context.
+  final AssetPickerTextDelegate? textDelegate;
+
+  /// Thumbnail size in the grid.
+  final ThumbnailSize gridThumbnailSize;
+
+  /// Preview thumbnail size in the crop viewer.
+  final ThumbnailSize? previewThumbnailSize;
+
+  /// {@macro wechat_assets_picker.PathNameBuilder}
+  final PathNameBuilder<AssetPathEntity>? pathNameBuilder;
+
+  /* [InstaAssetPickerBuilder] config */
+
+  /// Specifies the text title in the picker [AppBar].
+  final String? title;
+
+  /// Customize the display and export options of crops
+  final InstaAssetCropDelegate cropDelegate;
+
+  /// Specifies if the picker should be closed after assets selection confirmation.
+  ///
+  /// Defaults to `false`.
+  final bool closeOnComplete;
+
+  /// Specifies if the assets should be cropped when the picker is closed.
+  /// Set to `true` if you want to perform the crop yourself.
+  ///
+  /// Defaults to `false`.
+  final bool skipCropOnComplete;
+
+  /// The [Widget] to display on top of the assets grid view.
+  ///
+  /// Default is unselect all assets button.
+  final InstaPickerActionsBuilder? actionsBuilder;
+}
+
 class InstaAssetPicker {
   InstaAssetPickerBuilder? builder;
 
   void dispose() {
     builder?.dispose();
+    builder = null;
   }
 
   static AssetPickerTextDelegate defaultTextDelegate(BuildContext context) {
@@ -74,11 +173,11 @@ class InstaAssetPicker {
   /// Since the exception is thrown from the MethodChannel it cannot be caught by a try/catch
   ///
   /// check `AssetPickerDelegate.permissionCheck()` from flutter_wechat_assets_picker package for more information.
-  static Future<PermissionState> _permissionCheck() =>
+  static Future<PermissionState> _permissionCheck(RequestType? requestType) =>
       AssetPicker.permissionCheck(
-        requestOption: const PermissionRequestOption(
+        requestOption: PermissionRequestOption(
           androidPermission: AndroidPermission(
-            type: RequestType.image,
+            type: requestType ?? RequestType.common,
             mediaLocation: false,
           ),
         ),
@@ -87,11 +186,13 @@ class InstaAssetPicker {
   /// Open a [ScaffoldMessenger] describing the reason why the picker cannot be opened.
   static void _openErrorPermission(
     BuildContext context,
-    AssetPickerTextDelegate textDelegate,
+    AssetPickerTextDelegate? textDelegate,
     Function(BuildContext context, String error)? customHandler,
   ) {
+    final text = textDelegate ?? defaultTextDelegate(context);
+
     final defaultDescription =
-        '${textDelegate.unableToAccessAll}\n${textDelegate.goToSystemSettings}';
+        '${text.unableToAccessAll}\n${text.goToSystemSettings}';
 
     if (customHandler != null) {
       customHandler(context, defaultDescription);
@@ -108,6 +209,14 @@ class InstaAssetPicker {
   static ThemeData themeData(Color? themeColor, {bool light = false}) =>
       AssetPicker.themeData(themeColor, light: light);
 
+  static void _assertRequestType(RequestType requestType) {
+    assert(
+        requestType == RequestType.image ||
+            requestType == RequestType.video ||
+            requestType == RequestType.common,
+        'Only images and videos can be shown in the picker for now');
+  }
+
   /// When using `restorableAssetsPicker` function, the picker's state is preserved even after pop
   ///
   /// ⚠️ [InstaAssetPicker] and [provider] must be disposed manually
@@ -121,46 +230,16 @@ class InstaAssetPicker {
   /// Set [onPermissionDenied] to manually handle the denied permission error.
   /// The default behavior is to open a [ScaffoldMessenger].
   ///
-  /// Crop options
-  /// - Set [cropDelegate] to customize the display and export of crops.
-  ///
   /// Those arguments are used by [InstaAssetPickerBuilder]
   ///
-  /// - Set [provider] of type [DefaultAssetPickerProvider] to specifies picker options.
+  /// - Set [provider] getter of type [DefaultAssetPickerProvider] to specifies picker options.
+  /// Getter needed to initialize the provider state after permission check.
   /// This argument is required.
-  ///
-  /// - Set [gridCount] to specifies the number of assets in the cross axis.
-  /// Defaults to [_kGridCount], like instagram.
-  ///
-  /// - Set [pickerTheme] to specifies the theme to apply to the picker.
-  /// It is by default initialized with the `primaryColor` of the context theme.
-  ///
-  /// - Set [textDelegate] to specifies the language to apply to the picker.
-  /// Default is the locale language from the context.
-  ///
-  /// - Set [title] to specifies the text title in the picker [AppBar].
-  ///
-  /// - Set [closeOnComplete] to specifies if the picker should be closed
-  /// after assets selection confirmation.
   ///
   /// - The [onCompleted] callback is called when the assets selection is confirmed.
   /// It will as argument a [Stream] with exportation details [InstaAssetsExportDetails].
   ///
-  /// - Set [loadingIndicatorBuilder] to specifies the loader indicator
-  /// to display in the picker.
-  ///
-  /// - Set [limitedPermissionOverlayPredicate] to specifies if the limited
-  /// permission overlay should be displayed.
-  ///
-  /// - Set [specialItemPosition] to allows users to set a special item in the picker
-  /// with several positions. Since the grid view is reversed, [SpecialItemPosition.prepend]
-  /// will be at the top and [SpecialItemPosition.append] at the bottom.
-  /// Defaults to [SpecialItemPosition.none].
-  ///
-  /// - Set [specialItemBuilder] to specifies [Widget] for the the special item.
-  ///
-  /// - Set [actionsBuilder] function to specifies the [Widget]s to display
-  /// on top of the assets grid view. Default is unselect all assets button.
+  /// - Set [pickerConfig] to specifies more optional parameters for the picker.
   Future<List<AssetEntity>?> restorableAssetsPicker(
     BuildContext context, {
     Key? key,
@@ -169,57 +248,35 @@ class InstaAssetPicker {
     Function(BuildContext context, String delegateDescription)?
         onPermissionDenied,
 
-    /// Crop options
-    InstaAssetCropDelegate cropDelegate = const InstaAssetCropDelegate(),
-
-    /// InstaAssetPickerBuilder options
-    int gridCount = _kGridCount,
-    required DefaultAssetPickerProvider provider,
-    ThemeData? pickerTheme,
-    AssetPickerTextDelegate? textDelegate,
-    String? title,
-    bool closeOnComplete = false,
+    /// InstaAssetPickerBuilder parameters
+    required DefaultAssetPickerProvider Function() provider,
     required Function(Stream<InstaAssetsExportDetails> exportDetails)
         onCompleted,
-    Widget Function(BuildContext context, bool isAssetsEmpty)?
-        loadingIndicatorBuilder,
-    LimitedPermissionOverlayPredicate? limitedPermissionOverlayPredicate,
-    Widget? Function(BuildContext context, AssetPathEntity? path, int length)?
-        specialItemBuilder,
-    SpecialItemPosition? specialItemPosition,
-    InstaPickerActionsBuilder? actionsBuilder,
+    InstaAssetPickerConfig pickerConfig = const InstaAssetPickerConfig(),
   }) async {
-    assert(provider.requestType == RequestType.image,
-        'Only images can be shown in the picker for now');
-
-    final text = textDelegate ?? defaultTextDelegate(context);
-
     PermissionState? ps;
-    if (builder == null) {
-      try {
-        ps = await _permissionCheck();
-      } catch (e) {
-        _openErrorPermission(context, text, onPermissionDenied);
-      }
+    try {
+      ps = await _permissionCheck(null);
+    } catch (e) {
+      _openErrorPermission(
+        context,
+        pickerConfig.textDelegate,
+        onPermissionDenied,
+      );
+      return [];
     }
 
+    /// Provider must be initialized after permission check or gallery is empty (#43)
+    final restoredProvider = provider();
+    _assertRequestType(restoredProvider.requestType);
+
     builder ??= InstaAssetPickerBuilder(
-      initialPermission: ps ?? PermissionState.denied,
-      provider: provider,
-      title: title,
-      gridCount: gridCount,
-      pickerTheme: pickerTheme ?? themeData(Theme.of(context).primaryColor),
-      locale: Localizations.maybeLocaleOf(context),
+      initialPermission: ps,
+      provider: restoredProvider,
       keepScrollOffset: true,
-      textDelegate: text,
-      loadingIndicatorBuilder: loadingIndicatorBuilder,
-      limitedPermissionOverlayPredicate: limitedPermissionOverlayPredicate,
-      closeOnComplete: closeOnComplete,
-      cropDelegate: cropDelegate,
       onCompleted: onCompleted,
-      specialItemBuilder: specialItemBuilder,
-      specialItemPosition: specialItemPosition,
-      actionsBuilder: actionsBuilder,
+      config: pickerConfig,
+      locale: Localizations.maybeLocaleOf(context),
     );
 
     return AssetPicker.pickAssetsWithDelegate(
@@ -241,36 +298,12 @@ class InstaAssetPicker {
   /// Set [onPermissionDenied] to manually handle the denied permission error.
   /// The default behavior is to open a [ScaffoldMessenger].
   ///
-  /// Crop options
-  /// - Set [cropDelegate] to customize the display and export of crops.
-  ///
   /// Those arguments are used by [InstaAssetPickerBuilder]
-  ///
-  /// - Set [gridCount] to specifies the number of assets in the cross axis.
-  /// Defaults to [_kGridCount], like instagram.
-  ///
-  /// - Set [pickerTheme] to specifies the theme to apply to the picker.
-  /// It is by default initialized with the `primaryColor` of the context theme.
-  ///
-  /// - Set [textDelegate] to specifies the language to apply to the picker.
-  /// Default is the locale language from the context.
-  ///
-  /// - Set [title] to specifies the text title in the picker [AppBar].
-  ///
-  /// - Set [closeOnComplete] to specifies if the picker should be closed
-  /// after assets selection confirmation.
   ///
   /// - The [onCompleted] callback is called when the assets selection is confirmed.
   /// It will as argument a [Stream] with exportation details [InstaAssetsExportDetails].
   ///
-  /// - Set [loadingIndicatorBuilder] to specifies the loader indicator
-  /// to display in the picker.
-  ///
-  /// - Set [limitedPermissionOverlayPredicate] to specifies if the limited
-  /// permission overlay should be displayed.
-  ///
-  /// - Set [actionsBuilder] function to specifies the [Widget]s to display
-  /// on top of the assets grid view. Default is unselect all assets button.
+  /// - Set [pickerConfig] to specifies more optional parameters for the picker.
   ///
   /// Those arguments are used by [DefaultAssetPickerProvider]
   ///
@@ -298,12 +331,9 @@ class InstaAssetPicker {
   /// - Set [initializeDelayDuration] to specifies the delay before loading the assets
   /// Defaults to [_kInitializeDelayDuration].
   ///
-  /// - Set [specialItemPosition] to allows users to set a special item in the picker
-  /// with several positions. Since the grid view is reversed, [SpecialItemPosition.prepend]
-  /// will be at the top and [SpecialItemPosition.append] at the bottom.
-  /// Defaults to [SpecialItemPosition.none].
-  ///
-  /// - Set [specialItemBuilder] to specifies [Widget] for the the special item.
+  /// - Set [requestType] to specifies which type of asset to show in the picker.
+  /// Defaults is [RequestType.common]. Only [RequestType.image], [RequestType.common]
+  /// and [RequestType.common] are supported.
   static Future<List<AssetEntity>?> pickAssets(
     BuildContext context, {
     Key? key,
@@ -312,22 +342,12 @@ class InstaAssetPicker {
     Function(BuildContext context, String delegateDescription)?
         onPermissionDenied,
 
-    /// Crop options
-    InstaAssetCropDelegate cropDelegate = const InstaAssetCropDelegate(),
-
-    /// InstaAssetPickerBuilder options
-    int gridCount = _kGridCount,
-    ThemeData? pickerTheme,
-    AssetPickerTextDelegate? textDelegate,
-    String? title,
-    bool closeOnComplete = false,
+    /// InstaAssetPickerBuilder parameters
     required Function(Stream<InstaAssetsExportDetails> exportDetails)
         onCompleted,
-    Widget Function(BuildContext context, bool isAssetsEmpty)?
-        loadingIndicatorBuilder,
-    LimitedPermissionOverlayPredicate? limitedPermissionOverlayPredicate,
+    InstaAssetPickerConfig pickerConfig = const InstaAssetPickerConfig(),
 
-    /// DefaultAssetPickerProvider options
+    /// DefaultAssetPickerProvider parameters
     List<AssetEntity>? selectedAssets,
     int maxAssets = defaultMaxAssetsCount,
     int pageSize = defaultAssetsPerPage,
@@ -335,21 +355,22 @@ class InstaAssetPicker {
     SortPathDelegate<AssetPathEntity>? sortPathDelegate =
         SortPathDelegate.common,
     bool sortPathsByModifiedDate = false,
-    FilterOptionGroup? filterOptions,
+    PMFilter? filterOptions,
     Duration initializeDelayDuration = _kInitializeDelayDuration,
-    Widget? Function(BuildContext context, AssetPathEntity? path, int length)?
-        specialItemBuilder,
-    SpecialItemPosition? specialItemPosition,
-    InstaPickerActionsBuilder? actionsBuilder,
+    RequestType requestType = RequestType.common,
   }) async {
-    final text = textDelegate ?? defaultTextDelegate(context);
+    _assertRequestType(requestType);
 
     // must be called before initializing any picker provider to avoid `PlatformException(PERMISSION_REQUESTING)` type exception
     PermissionState? ps;
     try {
-      ps = await _permissionCheck();
+      ps = await _permissionCheck(requestType);
     } catch (e) {
-      _openErrorPermission(context, text, onPermissionDenied);
+      _openErrorPermission(
+        context,
+        pickerConfig.textDelegate,
+        onPermissionDenied,
+      );
       return [];
     }
 
@@ -358,7 +379,7 @@ class InstaAssetPicker {
       maxAssets: maxAssets,
       pageSize: pageSize,
       pathThumbnailSize: pathThumbnailSize,
-      requestType: RequestType.image,
+      requestType: requestType,
       sortPathDelegate: sortPathDelegate,
       sortPathsByModifiedDate: sortPathsByModifiedDate,
       filterOptions: filterOptions,
@@ -368,25 +389,21 @@ class InstaAssetPicker {
     final InstaAssetPickerBuilder builder = InstaAssetPickerBuilder(
       initialPermission: ps,
       provider: provider,
-      title: title,
-      gridCount: gridCount,
-      pickerTheme: pickerTheme ?? themeData(Theme.of(context).primaryColor),
-      locale: Localizations.maybeLocaleOf(context),
       keepScrollOffset: false,
-      textDelegate: text,
-      loadingIndicatorBuilder: loadingIndicatorBuilder,
-      limitedPermissionOverlayPredicate: limitedPermissionOverlayPredicate,
-      closeOnComplete: closeOnComplete,
-      cropDelegate: cropDelegate,
       onCompleted: onCompleted,
-      specialItemBuilder: specialItemBuilder,
-      specialItemPosition: specialItemPosition,
-      actionsBuilder: actionsBuilder,
+      config: pickerConfig,
+      locale: Localizations.maybeLocaleOf(context),
     );
 
     return AssetPicker.pickAssetsWithDelegate(
       context,
       delegate: builder,
+      permissionRequestOption: PermissionRequestOption(
+        androidPermission: AndroidPermission(
+          type: requestType,
+          mediaLocation: false,
+        ),
+      ),
       useRootNavigator: useRootNavigator,
       pageRouteBuilder: pageRouteBuilder,
     );